- Moving file structures around. Tests and export working
