+++ src/coin/baseCoin/enum.ts
@@ -1,0 +1,10 @@
+export enum Network {
+  Test,
+  Main,
+}
+
+export enum TransactionType {
+  Send,
+  Recieve,
+  WalletInitialization,
+}
