+++ src/coin/trx/signature.ts
@@ -1,4 +1,4 @@
-import { BaseSignature } from "../..";
+import { BaseSignature } from "../baseCoin/iface";
 
 export class Signature implements BaseSignature {
   signature: string;
